+++ src/diag.c
@@ -471,6 +471,7 @@ static const struct {
    { "Aggregates", { [STD_08] = "9.3.3", [STD_93] = "7.3.2" } },
    { "Interface object declarations", { [STD_08] = "6.5.2",
         [STD_93] = "4.3.2" } },
+   { "Association lists", { [STD_08] = "6.5.7", [STD_93] = "4.3.2.2" } },
 };
 
 diag_t *diag_new(diag_level_t level, const loc_t *loc)